+++ app/src/main/java/de/christian2003/chaching/plugin/presentation/view/transfers/TransfersScreen.kt
@@ -1,21 +1,33 @@
 package de.christian2003.chaching.plugin.presentation.view.transfers
 
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.calculateEndPadding
+import androidx.compose.foundation.layout.calculateStartPadding
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.windowInsetsBottomHeight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.TopAppBarScrollBehavior
+import androidx.compose.material3.TopAppBarState
+import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
 import androidx.compose.runtime.getValue
+import androidx.compose.ui.input.nestedscroll.nestedScroll
+import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringArrayResource
 import androidx.compose.ui.res.stringResource
@@ -26,6 +38,7 @@ import de.christian2003.chaching.domain.type.Type
 import de.christian2003.chaching.plugin.presentation.ui.composables.ConfirmDeleteDialog
 import de.christian2003.chaching.plugin.presentation.ui.composables.EmptyPlaceholder
 import de.christian2003.chaching.plugin.presentation.ui.composables.Headline
+import de.christian2003.chaching.plugin.presentation.ui.composables.NavigationBarProtection
 import de.christian2003.chaching.plugin.presentation.ui.composables.TransferListItem
 
 
@@ -45,6 +58,9 @@ fun TransfersScreen(
 ) {
     val types: List<Type> by viewModel.allTypes.collectAsState(emptyList())
     val transfers: List<Transfer> by viewModel.allTransfers.collectAsState(emptyList())
+    val appBarState: TopAppBarState = rememberTopAppBarState()
+    val scrollBehavior: TopAppBarScrollBehavior = TopAppBarDefaults.enterAlwaysScrollBehavior(appBarState)
+
     Scaffold(
         topBar = {
             TopAppBar(
@@ -60,14 +76,20 @@ fun TransfersScreen(
                             contentDescription = ""
                         )
                     }
-                }
+                },
+                scrollBehavior = scrollBehavior
             )
-        }
+        },
+        modifier = Modifier.nestedScroll(scrollBehavior.nestedScrollConnection)
     ) { innerPadding ->
         Column(
             modifier = Modifier
                 .fillMaxSize()
-                .padding(innerPadding)
+                .padding(
+                    start = innerPadding.calculateStartPadding(LocalLayoutDirection.current),
+                    top = innerPadding.calculateTopPadding(),
+                    end = innerPadding.calculateEndPadding(LocalLayoutDirection.current)
+                )
         ) {
             if (transfers.isEmpty()) {
                 EmptyPlaceholder(
@@ -88,10 +110,21 @@ fun TransfersScreen(
                     },
                     onQueryTransferType = { transfer ->
                         viewModel.getTypeForTransfer(transfer, types)
-                    }
+                    },
+                    windowInsets = WindowInsets(
+                        bottom = innerPadding.calculateBottomPadding()
+                    )
                 )
             }
         }
+
+        NavigationBarProtection(
+            color = MaterialTheme.colorScheme.surfaceContainer.copy(0.5f),
+            windowInsets = WindowInsets(
+                bottom = innerPadding.calculateBottomPadding()
+            )
+        )
+
         if (viewModel.transferToDelete != null) {
             ConfirmDeleteDialog(
                 text = stringResource(
@@ -117,13 +150,15 @@ fun TransfersScreen(
  * @param onEditTransfer        Callback invoked to edit a transfer.
  * @param onDeleteTransfer      Callback invoked to delete a transfer.
  * @param onQueryTransferType   Callback invoked to query the type for a transfer.
+ * @param windowInsets          Insets for the screen.
  */
 @Composable
 private fun TransferList(
     transfers: List<Transfer>,
     onEditTransfer: (Transfer) -> Unit,
     onDeleteTransfer: (Transfer) -> Unit,
-    onQueryTransferType: (Transfer) -> Type?
+    onQueryTransferType: (Transfer) -> Type?,
+    windowInsets: WindowInsets
 ) {
     val groupedTransfers = transfers.groupBy { transfer ->
         transfer.valueDate.withDayOfMonth(1)
@@ -146,6 +181,11 @@ private fun TransferList(
                     onQueryTransferType = onQueryTransferType
                 )
             }
+        }
+        item {
+            Box(
+                modifier = Modifier.windowInsetsBottomHeight(windowInsets)
+            )
         }
     }
 }